pyproject.toml
[tool.poetry]
name = "sfb"
version = "0.1.4"
description = "SQL tester and cost estimator for Google BigQuery"
authors = ["Toshifumi Tsutsumi <rev.to12@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.6, <3.10"
google-cloud-bigquery = "^2.13.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
mypy = "^0.812"
black = "^20.8b1"
pytest-cov = "^2.11.1"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
